# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Bridge API
  version: 1.125.0
  description: 'Bridge — federated, self-hosted communication platform REST API.


    **Authentication:** Tüm korumalı endpoint''ler `Authorization: Bearer <accessToken>` header''ı gerektirir.

    Access token''lar kısa ömürlüdür (15 dk). `POST /api/refresh` ile yenilenebilir (httpOnly cookie).


    **Rate Limiting:** Endpoint başına granüler rate limit uygulanır. 429 döndüğünde `Retry-After` header''ını kontrol edin.


    **Federation:** Bridge sunucular arası ActivityPub protokolü kullanır. Federation endpoint''leri `/api/federation` altındadır.

    '
  contact:
    name: Bridge Contributors
    url: https://github.com/bridge-app/bridge
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: https://your-bridge-instance.com/api
  description: Production
- url: http://localhost:3001/api
  description: Local development
...
paths:
...
  /interactions:
    post:
      tags:
      - Bots
      summary: Bot interaction webhook alici
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: integer
                data:
                  type: object
                token:
                  type: string
      responses:
        200:
          description: Interaction islendi
        401:
          $ref: '#/components/responses/Unauthorized'
      operationId: post_interactions
  /interactions/{interactionId}/callback:
    post:
      tags:
      - Bots
      summary: Interaction callback gonder
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: interactionId
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: integer
                data:
                  type: object
      responses:
        200:
          description: Callback islendi
      operationId: post_interactions_interactionId_callback
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. POST /api/login ile alınır, süresi 15 dakikadır.
    botToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bot token: `Authorization: Bot brg_bot_…` (sunucu sahibi bot oluştururken bir kez gösterilir).'
...
  responses:
...
    Unauthorized:
      description: Token eksik veya geçersiz
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
security:
- bearerAuth: []